Strengthen the atomic ordering requirement for removing a node from the kernel heap.
This might fix issue #1, since the problem seemed to be that the node was being removed
even though a new reference to it had just been made. An atomic CAS operation was
supposed to prevent that. Upgraded the `success` ordering to `SeqCst` to prevent it
being reordered around checks involving a different atomic variable.