Start refining the system call interface.
I've added the libraries `libphoenix` and `libdriver` to abstract
the system calls. There are currently no asynchronous system calls
(we don't have the infrastructure for it yet), but the userspace
API for them is starting to take shape. The API for interacting
with threads is also a little higher-level now than it was.