Finish the userspace side of a skeleton driver.
The virtio-gpu driver can now be cross-compiled as an executable
file entirely through Rust, and a new macro in `libphoenix` makes
it nearly seamless. Also, the userspace API for the `Device_Claim`
system call is set up. Once we implement the system call in the
kernel, we can test that it works and start actually implementing
the driver.