Add a global allocator to libphoenix.
As part of doing this, I had to work around an apparent bug in the
nightly Rust compiler by defining the panic handler in the root
crate through a macro.

Also, this global allocator is terrible as a global allocator. It
should be replaced very soon. This is only a temporary patch to allow
storing futures on the heap, in preparation for asynchronous system
calls.