Fix a bug in the page table logic regarding lazy page mapping.
When we mapped a page in a file without reading it into physical
memory, we asked the page tables to map it at physical address 0.
Later, we would expect to see a physical address of 0 when the
page needed to become resident in RAM. This worked when lazily
mapping a single page at a time, but if a second page was mapped as
part of the same operation, the physical address would be
incremented by a page, and so our check later on would fail.

This commit makes the page tables' `map` functions aware of the
possibility of being given no physical address, as opposed to a
physical address of 0.