Add more nuance to the return type of `ExecImage::load_segment_piece`.
This function used to return `Ok(None)` to indicate that another CPU
was using the file and the caller should switch to another thread
while it waits. But it also (accidentally) used the same return
value to indicate that the pages were successfully mapped as CoW.
Now, we properly distinguish between those outcomes.