Make the kernel able to accept strings from userspace.
This implementation is perhaps more complex than it needs to be:
we could just copy the string to kernelspace and access it there.
But I think the time we save by accessing the string by reference
is worth the complexity. We can also avoid checking for buffer
overflows and running out of memory because we access the string
by reference.

One reason for the complexity is that we don't pause the world
while we use the string, and so it could be changed while we're
reading it. Therefore, a userspace string is accessed as a stream
rather than as a random-access array. This works well for comparing
the value to a known string, such as when searching the device
tree.

Another reason is that the kernel and the process are mapped in
different address spaces. So we need to convert the userspace
address to a physical address before we can find the data in the
kernel's address space.

No well-behaved program will change a string between giving it to
the kernel and receiving a response, but we need to guard against
ill-behaved programs that do. For instance, a malicious program
might quickly change an input string to fool the kernel's
permissions checks (which are not yet implemented). For that
reason, assuming that two bytes that are read from the same
userspace address will ever have the same value, even if that
address was provided as an argument to a system call, is *undefined
behavior*. We must avoid that assumption at all times.