Begin the implementation of the Device_Claim system call.
Right now, we can walk the device tree to find the requested device
and then atomically mark it as claimed. The next step is to map its
MMIO resources to the process's address space, and then we can
construct an object to represent it and return that object to
userspace.

Note that, although everything described so far is synchronous in
nature, this system call will need an asynchronous API when we
implement asking the user for permission. To that end, this commit
also includes updates to libphoenix's `future` module that allow
a clean API.