Add new system calls for allocating and freeing memory.
There are two options for allocating memory: `memory_alloc` and `memory_alloc_phys`. The latter
guarantees an upper bound on the physical address that's used (since not all devices have 32 or
64 address lines) and also guarantees that the memory will remain resident in memory. (Note that
we don't have any code enforcing that second guarantee yet. We'll have to revisit this when we
implement page-swapping.)

The `memory_free` system call doesn't actually free any memory yet, since we're not yet keeping
track of which blocks each process owns. (We're not even keeping track of processes themselves
yet.) But the API is there, so we can write a global userspace allocator based on `memory_alloc`
and `memory_free`.