Replace the userspace global bump allocator with an allocator that gets pages from the kernel.
The new allocator still needs work. Specifically, it currently naively assumes that every
allocation or deallocation should correspond with a system call. This wastes a lot of memory
because the kernel can't hand out less than a page at a time, and it also has a really high
latency and low throughput. We can fix both problems by having the allocator ask the kernel
for a large buffer and then use that buffer until it's full.

The buffer should be managed by either a bump allocator or a slab allocator. I'm not sure
right now which would be better.