Add virtqueues to the virtio driver.
We now have a `VirtQueue` type (and also private types for its various parts)
that allows initialization and an asynchronous `send_recv` operation. We also
have a `ResponseFuture` type that represents the result of `send_recv`. This
is all untested (can't really test it without a working driver that uses it),
but it was ported from `gfx_prototype.rs`, with only minor changes, so it
_should_ work just fine.

We still have to finish the device initialization in
`virtio/src/lib.rs:init_mmio`, and we need an executor (and an associated
waker type) to poll the `ResponseFuture`s. The executor should probably be in
its own crate.