Make an executor for the futures produced by virtqueues.
This executor is actually a very generic one, so we will probably want to
move it to its own crate later. For now, though, VirtIO drivers are the
only programs that use it, so it can stay in the virtio crate.

It compiles, but it's otherwise untested, so use it with caution.