Allow padding bytes in the kernel's allocator.
We used to require the number of bytes being allocated to be a
multiple of the size of `T`, but that got in the way of allocating
a whole page to store something smaller than a page. Now we just
require it to be at least as large as `T`. This disallows
allocating an empty array, but we probably don't want to do that
anyway.