Optimize the InitRD's build script.
We were reading the whole files and encoding them in byte arrays, which
took a somewhat long time and, more importantly, eventually led to a
source code file long enough that it crashed rustc (or, at least, that
seems to have been the cause). Now we use `include_bytes!` and let
rustc turn them into byte arrays for us, which seems to work better.