libphoenix: Make syscall futures allocation-free.
We originally avoided dynamically sized types by allocating a `Vec` for
each syscall return value. Not only was this inefficient, but it
actually led to a deadlock every time. Allocating memory required its
own system call, and this loop would repeat until all the memory was
exhausted. We now handle the return value in place, using a DST.