virtio-gpu driver: First iteration.
This commit lays a lot of the groundwork for the VirtIO GPU driver.
We now have a framework for adding more VirtIO commands and
encapsulating them in functions that will later be attached to the
driver's public API. For now, we have proof of correct operation in
that the driver successfully retrieves the display info (i.e. the
list of scanouts and their properties) from the device.