libphoenix: Add a sized constructor to `PhysBox`.
It was possible to construct a `PhysBox` by calling one of the physical
memory allocation functions on `Allocator`, and while that route is
necessary for most usage of DSTs, it's far from ergonomic. We now have
`PhysBox::new`, which works just like `Box::new`.

While `PhysBox::new` requires its argument to be `Sized`, it can also
be used for DSTs when combined with the implementation of `CoerceUnsized`.
For instance, this will work:

```
let x: PhysBox<[u8]> = PhysBox::new([0u8; 256]);
```