kernel: Avoid undefined behavior and incorrect atomic orderings in the heap.
I'd previously written some pointer casts to implement interior mutability on
an immutable reference to `self` in `MasterBlock::try_free`. This caused
instant undefined behavior, as the Rust compiler was free to assume that `*self`
would never be modified. (For instance, it could have cached its fields in
registers and not reread them when needed.) Now, we do the right thing by using
a `RefCell`.

There was also an atomic operation with some ordering arguments that, in
retrospect, I couldn't prove to myself were strict enough. I changed them to
stricter versions just in case.