kernel: Validate each individual allocation in the heap unit tests.
We already had a function to do this, but we were only using it in
the `many_threads` test. Now, we validate every allocation.
Specifically, we verify that each node is in the list right after
it's allocated and also that it is completely contained within a
master block. (Alignment is not checked.)