kernel: Streamline the heap node's reference to its master block.
We used to allow `None` as an alternative to a reference to the
static master block. This was necessary for the very first guard
that was made, since it had to be constructed during its master
block's construction--and, therefore, before a reference to that
block could be made. Since we no longer require the heap to contain
at least one node at all times, this workaround can be easily
removed.