libphoenix: Add basic profiling support.
The functionality is locked behind the `profiler` feature, although
the API is still available without that in the form of stubs (for
ergonomics at the call sites).

In the current form, the profiler consists of statically allocated
probes that count the number of times execution reaches each of
them. Combined with any way to keep track of time, this allows a
direct measurement of throughput. Probes can also be linked together
for the purpose of measuring latency (which, by Little's Law, is
calculable as the average number of threads visiting divided by the
throughput).

The next step is to add Coz-like experimentation, which will require
a way to select a particular probe and a way to slow down all other
probes in a controlled and measurable manner.