Kernel: Add the `time_now_unix` system call.
This system call returns the current time as a 64-bit UNIX timestamp.
(As a result, it is limited to whole-second precision.)

The kernel's implementation also provides for non-64-bit word sizes
by allowing the full value to be returned in smaller pieces by
multiple calls, but libphoenix currently only supports 64-bit (and,
in fact, will generate a compiler error if given a different word
size). We will need to fix this when porting to, for instance, 32-
bit ARM architectures.

The reason for using multiple calls instead of returning a future is
that retrieving the current time is often performance-critical
(because it's used in benchmarks). Allocation of a new future would
be too slow.