Kernel: Remove the `reserve` and `reserve_mut` allocator functions.
We never actually used them, and we won't ever need to use them in
the future. Here's why:

First, they were only usable for static RAM. Not MMIO, just RAM. But
there is no static RAM in the system that we can't completely account
for at compile time. Second, even if we did use the reserve functions,
they'd be unreliable after the first call to `malloc`, since that
might have already claimed the memory. So we would have to invoke them
before the allocator had finished its initialization.

We do currently reserve some memory during the allocator's
initialization, but instead of using these functions, we just remove
the memory from the memory map entirely. This must likewise be done
before allocator initialization finishes, but it can also be done
before it starts.