Kernel: Remove the `memory::phys::block::Block` type.
After removing `reserve` in the previous commit, `Block` was only
used in one place, and its job was easily replaceable by a simple
byte slice. In general, there is never a situation where the kernel
needs to allocate memory through the heap without also needing the
ability to modify it.

The original idea was that the kernel would allocate, for instance,
the x86 architecture's BDA. But for reasons described in commit
822a1834919fec4c3fb45e33fd5c11a435fb2250, there is no reason to use
the heap when doing this. Instead, we can modify the memory map and
use static slices and pointers.