Kernel: Add profiler output for the bootloader.
It's pretty barebones right now, but it's almost in a state where
it will be useful. In particular, we'll have to add a latency
measurement and use a more precise timer. (The 1-second resolution
of the PL031 means we will currently only be able to see about 20%
or better improvements in boot speed, since it currently takes 5
seconds.)