build tools: Eliminate unwarranted rebuilds of the InitRD.
In order to do this, we build the InitRD for every supported target at once and no longer rebuild when the
`PHOENIX_TARGET` environment variable changes. We still rebuild when the InitRD contents change, but that
environment variable dependency was somehow confusing Cargo, making it think we always needed to rerun the build
script.