idl2rust: Write an IDL-to-Rust transpiler (incomplete).
The transpiler currently supports interfaces, dictionaries, constants, operations, attributes, callback
interfaces and functions, and most types (including unions).

It does not yet support mixins; iterables; maplike and setlike types; buffer types; or array types.

Support for partial interfaces and dictionaries is not planned. It would be hard to implement and, outside the
context of a prose specification, would just make the IDL harder to read.