idl2rust: Implement several no-op extended attributes.
Specifically, this commit adds support for these extended attributes:
 * `[Clamp]`
 * `[EnforceRange]`
 * `[Exposed]`
 * `[SameObject]`
 * `[Unscopable]`
 * `[LegacyNullToEmptyString]`
 * `[LegacyTreatNonObjectAsNull]`
 * `[LegacyUnenumerableNamedProperties]`
 * `[LegacyUnforgeable]`

Since these extended attributes (with the exception of `SameObject`) are concerned only with conversion of
values from ECMAScript, they are implemented as no-ops. The only effect is to suppress the warning about an
unrecognized extended attribute.

In the case of `SameObject`, it is still a no-op because the additional requirement it places on a read-only
attribute (the only place where it's legal) was already consistent with how we did the Rust binding (using a
`&self` function instead of a `&mut self` function). We can't use a constant or a static variable because the
attribute doesn't require different objects to return the same value from that attribute.