idl2rust: Add support for custom extended attributes.
Now, instead of warning about an unrecognized extended attribute, the transpiler generates a call to a
client-defined macro that is free to modify the generated tokens in any way. This is necessary for separating
the IDL parser from the DOM and any other systems that might use IDL, since the IDL standard doesn't define all
the extended attributes that exist.