idl2rust: Remove the camel-case to snake-case conversion of identifiers.
The conversion was unnecessary and just made reasoning about both the parser's internals and the generated code
more difficult. Forcing standard names to conform to Rust's standard naming conventions is not worth the
trouble.

I also fixed a few minor errors in the crate's doc comment that were unrelated to the case conversion.