idl2rust: Use prefixes instead of suffixes to mark overloaded identifiers.
The old system wasn't robust. Using the example in the crate's rustdoc comment, if in the IDL fragment we
defined two `stroke` methods and a `strokeO1` method, the suffix approach would lead to a compiler error by
trying to redefine `strokeO1`. The prefix method, which begins with an underscore, ensures that no valid IDL
identifier will conflict with the Rust identifiers we use to support overloading.