libphoenix: Fix a race condition in the profiler.
Each probe was marked as registered before its initialization was complete. This prevented two threads
from attempting to initialize the same probe simultaneously, but it made possible a situation wherein
the second thread thinks the probe is initialized and logs its visit. This would have caused the
calculations to be way off in the worst cases.

We now block the second thread until initialization is actually complete.