Kernel+libphoenix+virtio-gpu: Make all system calls synchronous.
This is a major design change that really couldn't be done in smaller pieces. Still, there are two parts to it:

1. All the code that allocates, initializes, and uses futures containing the return values of system calls has been
removed.

2. The system call ABI has been updated to allow two `usize`-sized return values instead of just one. This was necessary
for `memory_alloc_phys`, since that system call needs to return two addresses. This can easily be extended later to
allow even more return values, but two are enough for now.

Simply by avoiding allocating futures (which was, to be fair, being done in an extremely inefficient way), this commit
reduced the runtime of the `virtio-gpu` graphics test by about 75%!