Kernel: When accessing userspace memory from the kernel, load it from the file if necessary.
We'd already implemented loading pages lazily from the executable file in response to the userspace application's
attempts to access them, but we were simply reporting the pages as missing when the kernel tried to access a page that
hadn't been loaded yet. This was a problem because it prevented the kernel from reading some strings when they were
provided as system call arguments, and it surely would have led to more problems in the future.