Kernel: Add slab allocators for page-sized allocations.
Although a single slab allocator runs in constant time, we string a bunch of slab allocators together in a linked list
for maximum flexibility. A new slab allocator is added to the head of the list whenever the existing ones are out of
pages to give. The result of all this is that memory allocations take O(n) time (where n is the number of previous
allocations) regardless of whether we use slab allocators or not.

The real benefit is that a slab allocator has far less computational overhead than the linked-list heap. Additionally,
the most recently created slab allocator, which is also the most likely not to be completely full, is checked first when
trying to allocate a new page. The original linked-list heap still exists, though, in order to handle allocations of any
size or at specified physical addresses. In fact, the slab allocators are themselves allocated from that heap.

The slab allocator implementation can support many other sizes in theory, just by changing a generic constant. But I
don't think they have enough benefits to actually use other sizes. I plan to add bump allocators to improve the
efficiency of smaller allocations, and larger allocations are rare enough that the linked-list heap should handle them
just fine.