virtio: Fix the checks done before deciding to notify the device about a new buffer.
According to the spec, the `flags` member of the used ring must be ignored when
VIRTIO_F_EVENT_IDX is negotiated, and a memory barrier is required before reading either
`flags` or `avail_event`. We were always using the `flags` member, and we didn't use a
memory barrier. (We used volatile reads, but those only affect reorderings by the
compiler, not reorderings by the CPU.) Now both of these requirements are met.