virtio: Fix the `next` member of the last descriptor in the descriptor table.
The spec says that, if the `VIRTIO_F_IN_ORDER` feature is negotiated, the last descriptor
must have `next` set to 0 when made available to the device. But even without that
requirement, it's likely that we have assumed somewhere (or will in the future) that the
unused descriptors always form a circular linked list.