virtio: Prioritize handling used buffers in order over handling them out of order.
Before, if `VIRTIO_F_IN_ORDER` was negotiated and the device signaled that it had used
both buffers A and B, we might not have handled buffer A at all. If buffer A's future
was asleep, buffer B would see that its index came up in the list and use the normal
out-of-order logic to handle the buffer and wake a future for the next one. Now, we
check the feature flag first, so buffer B's future knows to wake buffer A's future
before returning its own descriptors to the table.