virtio: Make descriptors in the same chain contiguous when using `VIRTIO_F_IN_ORDER`.
We normally have to allocate one descriptor at a time, since the linked list of free
descriptors may get jumbled up over time as the device uses them out of order. But
with the `VIRTIO_F_IN_ORDER` feature negotiated, the device is required to use and
return them in order, which allows us to allocate them all at once. Moreover, the
spec requires us to keep them all contiguous in the descriptor table, so we're now
more spec-compliant.