virtio: Fix a data race when adding entries to the driver ring.
It may never have happened in practice, but it was theoretically possible for a data
race to occur in `DriverRing::set_next_entry` as follows:

1. Thread A adds an entry, reveals it to the device by calling `add_idx`, and then is
   pre-empted by the OS.
2. Thread B adds an entry, reveals it to the device by calling `add_idx`, and continues
   on its way. Thread A is still going to look at this entry when it resumes.
3. More threads add more entries. Eventually, thread C starts adding an entry at the same
   index that thread A originally used. (The VirtIO specification has no built-in
   protection against this, beyond a guarantee that no entry can be overwritten before it
   is revealed to the device.) Thread C is pre-empted by the OS after it claims the index
   but before it can set the descriptor ID in the entry.
4. Thread D adds an entry at the same index that thread B originally used--the same
   index that thread A will inspect. Seeing that thread C hasn't finished its own work,
   thread D sets the "updated" flag and returns.
5. Thread A resumes execution, sees thread D's "updated" flag, and calls `add_idx` again.
   It intends to reveal thread D's entry to the device, but it actually reveals thread C's
   entry. The device accesses whatever physical address happened to be in the entry
   already, and undefined behavior ensues.

The data race has been resolved by giving each entry a mutex alongside its "updated" flag
so that, during the critical section between incrementing the ring's `idx` field and
checking whether the next entry has been updated, no new entries can be added that
overwrite that next entry. So threads C and D in the story above will wait until thread A
is certain about whether it should continue or not. New entries can still be added at other
indices during this time, and overwriting other entries is fine as long as it obeys the
mutex and the specification.