virtio: Delay notification of the device until the appropriate buffer is actually revealed.
If the `VIRTIO_F_EVENT_IDX` feature is negotiated, the device wants to be notified after
a buffer at a particular index is made available. Spurrious notifications are okay, but
missed notifications are not. We were sending a notification as soon as `send_recv` found
that we'd used the right index, regardless of whether the buffer had been made available
or not--and we weren't sending it again after making the buffer available. Now we send a
notification when the buffer is actually made available, even if this task wasn't
responsible for putting it in the driver ring.