virtio: Use the right index when setting up a descriptor to send to the device.
The bug stemmed from using a slice of indices into the table. I'd accidentally used
indices for the slice as indices for the table. In order to prevent this in the
future, the descriptor table and the driver ring are now indexed by `u16` rather
than by `usize`. This is generally more convenient too, as it means fewer typecasts.