virtio: Honor the `VIRTIO_F_IN_ORDER` feature by keeping descriptors in order when freed.
We were putting the descriptors at the head of the free list, and we still do that when
the feature hasn't been negotiated. But that was incorrect when multiple descriptor
chains were allocated before the first one was freed. Instead, the descriptors should
be added back to the list at the end.

Fortunately, the fix is simple: if the descriptors must remain in order, do nothing. They
are already at the end of the list of free descriptors by virtue of comprising the oldest
chain that is still allocated.