virtio-input: Await events in parallel rather than sequentially.
The previous version of the code used futures, but only one per device. The result was
that input would work for about 100 events and then lock up, waiting for the wrong buffer.

Given the API currently provided by the `virtio` crate, we needed to have one future per
buffer instead.