libphoenix: Expose the system calls in a C-like static library.
This exposed a potential bug in `thread_spawn`: we were passing it a `fn()`, converting
that to a `usize` to send it to the kernel, and then calling it from the kernel as if it
used the C ABI. But, of course, it uses the Rust ABI instead. Even for a function that
takes no arguments and returns nothing, we shouldn't assume anything about the Rust ABI,
since it is, as far as I'm aware, permanently unstable.

So now we send an `extern "C" fn(usize)` to the kernel. The argument is there to avoid
forcing client code to use `extern "C"` for all of its thread entry points. It is equal
to the address of the function (of type `fn()`) that the client provides. The function
that the kernel calls is a stub whose purpose is to switch to the Rust ABI by calling
the client's function.

The version of `thread_spawn` exported from the static library just uses the C ABI and
allows the client to specify an arbitrary argument. This conveniently matches the
requirements of the POSIX function `pthread_create`! :)