libc: Change `WEXITSTATUS` to make `EXIT_FAILURE` representable.
`EXIT_FAILURE` is currently defined as -1, but the previous definition of `WEXITSTATUS`
always produced a positive number and would have interpreted that exit code as 255. We
now interpret the low 8 bits as a signed integer instead.

This is likely irrelevant for POSIX conformance, since POSIX is silent on the actual
data type produced by `WEXITSTATUS`, saying only that it evaluates to the lowest 8
bits of the status code. Any application seeking maximum portability should use a
typecast to check `(uint8_t)WEXITSTATUS(status) == (uint8_t)EXIT_FAILURE`.