libc: Make the layout of `struct sigaction` the same in all versions of C.
Since C90 doesn't support anonymous unions, I'd used the preprocessor to turn it
into two separate fields in that case. But that breaks binary compatibility with
the compiled library. So now we just always use separate fields, foregoing the
opportunity to save a little memory by using a union. Given the small number of
`sigaction` objects that any given process will use, the larger memory footprint
is not an issue.