libphoenix: Enable compilation of the Rust library without also compiling the C library.
This is difficult because Cargo compiles all the crate types listed in Cargo.toml at the
same time in order to minimize the work involved. By adding a "clib" subdirectory filled
with symlinks to the contents of the root, we can make a new Cargo.toml that builds the
C library from the same source code, freeing the main Cargo.toml to compile only the
Rust library.

The reason we have to do this is that the "staticlib" crate type requires the
`#[panic_handler]` and `#[alloc_error_handler]` lang items to be defined, whereas the
Rust library needs them to be defined only once in the entire dependency graph. And the
kernel, which links to the Rust library, defines these lang items itself. Now Cargo will
stop complaining about the lang items being undefined in the C library when compiling a
crate that uses only the Rust library.