kernel: Add the concept of processes.
Rather strangely, a process is owned by its threads rather than the other way
around. This allows a thread to very easily query its process for information,
and it ensures that the process is cleaned up once the last thread finishes.

This commit also does a much-needed refactoring of the scheduler, putting the
`Thread` and related types into their own module, separate from the main logic.