virtio-gpu: Remove a code pattern that relied on undefined behavior.
I'd used `mem::forget(mem::replace(...))` many times to initialize values without dropping
them. This works, in the sense that it writes a new value and doesn't drop the old one, but
it requires making a mutable reference to the thing being replaced. In this context, that
means a reference to uninitialized data, which is instant undefined behavior--even if the
value being replaced is plain-old-data. And since the optimizer is free to assume that
undefined behavior just doesn't happen, every code path that used this pattern could be
optimized away.

That never happened, but let's not give the optimizer any funny ideas.

The new pattern uses `MaybeUninit` and raw pointers to avoid ever making a reference. Raw
pointers are much more lenient; they just need to be properly aligned and point to allocated
objects. The objects don't even need to be of the right types, which lets us transmute as
necessary in the course of initialization.