libphoenix: Set the maximum address in `PhysBox::new` according to the size of `usize`.
We used the size of `*mut T` before. Since `T` implements `Sized`, this is currently the same
as the size of `usize`, but it might not be forever. Rust's Strict Provenance Experiment may
end up decoupling the concepts of pointer and address, such that a pointer can be larger than
an address. We just want the size of an address here, and `usize` is perfect for that. We also
already store the physical address as a `usize` anyway.