kernel: Return blocks of `MaybeUninit<u8>` from `malloc[_low]`.
The change from blocks of `u8` has two main benefits in terms of
statically enforcing invariants.

First, it reduces the appeal of the `mem::forget(mem::replace(...))`
antipattern that, in the case of uninitialized data, always produces
undefined behavior.

Second, it serves as a reminder when allocating memory for userspace
that the memory should be scrubbed first.