libc: Adhere to POSIX namespace rules when catering to C90 and C99.
POSIX defines requirements for the names of symbols exposed by the
standard headers for when the implementation needs to define some
nonstandard symbols. We were defining `restrict` in these headers, which
was not allowed. And although we undefined it in most (I'd forgotten
that step in signal.h), a POSIX-conforming application should be able to
expect that any macro defined before the headers are included, as long
as POSIX doesn't allow the headers to define a symbol with that name,
should remain the same after the headers. With this change, that
hypothetical assumption is true regardless of whether the version of C
being used has the `restrict` keyword or not.