crypt: Implement the SHA-512 algorithm.
This library isn't needed for anything at the moment, but the plan is to
use it for hashing passwords and file contents, both for security
purposes. I've spent no time optimizing it, as it seems fast enough for
the planned use cases.

I've also included two simple unit tests and three more complex ones
that use official test vectors from the NIST. Everything seems to work!