kernel: Change how `time_unix_now[_nanos]` reacts to invalid input.
It used to assume that `TimeSelector::Now` was the appropriate value to
use when it was given an invalid time selector. That would make it quite
hard to debug any errors that might arise from introducing a new variant
of `TimeSelector`, since an application running on an older version of
the kernel would get the wrong time back. Now it will at least get the
perhaps obviously wrong value of 0.