idl2rust: Start overhauling the parser.
The new parser has better static typing while being more declarative,
less procedural, and less object-oriented. The parser itself is
finished, as it results in a complete AST, but the final transpilation
step still needs work. I've left significant parts of the old parser in
fixme comments so I can remember how the transpilation was done before
if I need to.

Although the parser itself is less object-oriented, the transpilation
allows a convenience of OOP that is, by design, missing in Rust.
Specifically, if B inherits from A, then B inherits all of the behaviors
of A except those which are explicitly overridden. With the new `_super`
method included in every generated trait, each type that implements B
can specify an object that implements A and defines all the behaviors
that this type should inherit.