idl2rust: Parse extended attributes into a usable form.
They are no longer parsed exactly according to the grammar in the
specification, but that won't matter in practice. We will be able to
parse any extended attribute of any of the six types that are
specifically defined and will only fail to parse extended attributes
that the transpiler wouldn't know what to do with anyway.