kernel: Replace `drop` with `let _ =` to avoid UB.
Because `root` is a mutable reference, we can't just ignore its
existence while we create another pointer to the same memory, as that
would be UB. We also can't call `drop(root)` to end its lifetime because
dropping a reference has no effect. (The latest version of rustc rejects
it.) But we can move it without binding it, and that does end its
lifetime.