libc: Make sys/types.h both POSIX-compliant and simpler.
The POSIX-compliance comes from matching the signedness requirements in
the spec and making `dev_t` an integer type. (Unfortunately, that means
it can't store a UUID, since 128-bit integers are non-standard and, in
GCC, unsupported on 32-bit platforms.)

The simplicity comes from making `off_t` always 64-bit. It actually was
already always 64-bit, since we were testing for `INT64_MAX`, which is
always defined regardless of the platform.