libc: Simplify the `U*_MAX` definitions.
They are now defined as unsigned versions of -1, since the C standard
specifies that converting from a negative number to an unsigned number
is done by repeatedly adding 1 more than the maximum value of the
unsigned type. So -1 will always be equivalent to the maximum value,
regardless of its bit pattern.