libc: Make the I/O functions aware of stream orientation.
POSIX doesn't say exactly what should happen when the stream's
orientation is violated. It just says, "Byte input/output functions
cannot be applied to a wide-oriented stream, and wide-character
input/output functions cannot be applied to a byte-oriented stream." We
could read that as undefined behavior, or as a requirement for the
functions to fail in that situation. glibc does fail, but almost
silently: it doesn't set the stream's error flag, nor does it set errno.
I've decided to treat it as a real error, so the error flag is set, and
errno is set to EINVAL.

Note that this breaks `fputwc`'s current implementation, since it uses
the byte-oriented `_PHOENIX_fwrite_unlocked` to write the bytes. This
will need to be fixed somehow (ideally without duplicating the whole
`fwrite` implementation).