libc: Define `EWOULDBLOCK` as `EAGAIN`.
POSIX allows these two error codes to be identical, and whenever it
lists the errors defined for a function, it seems to always give
`EAGAIN` as an alternative to `EWOULDBLOCK`. I don't want to choose
`EWOULDBLOCK`, forget I'd done that, and later assume that I could just
check for `EAGAIN`.