libc: Make all I/O functions use `unsigned char` instead of `char`.
The `read` and `write` functions already did, but the functions in
`stdio.h` didn't consistently use that type. This was mainly because the
standard API uses `char*` as the types of arguments in functions like
`fputs` and `setvbuf`. But even though the user can provide a buffer of
type `char[]`, the C standard guarantees that the actual bytes making up
any object can be accessed as an array of type `unsigned char[]` in
exactly the way one would expect. Moreover, POSIX says that the contents
of the buffer at any given time are unspecified, so we don't have to
ensure that the bit pattern for each byte is a valid `char`.

There was also one instance of accidentally relying on a little-endian
`int` representation. That has also been fixed.