libc: Banish wide character and wide string literals from the codebase!
Practically speaking, there's never going to be an issue with using
them, but theoretically speaking, they're dangerous. The C standard
compels the compiler to convert them into `wchar_t` values as if by
`mbstowcs`, but it can't use the version of that function defined in our
libc. So we can't guarantee that our `wcstombs` is the inverse function
of whatever the compiler does.

So, instead, we will make as little use of wide characters as possible
and, when we must use them (e.g. for `perror`), we will do the
conversion from whatever encoding we use ourselves. In most cases
(printable ASCII characters), we can use POSIX's requirement that they
have the same values when stored in `wchar_t` as when stored in `char`
to avoid any runtime overhead.