libc: Remove all standard symbols and macro names from macro expansions.
As described at
[https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_xsh_chap02.html#tag_22_02_01],
macros must not expand to standard symbols that the application is
allowed to use as local identifiers (e.g. local variable names), since
the preprocessor knows nothing about scopes. The solution is to use
aliases with names that conforming applications aren't allowed to use,
such as those in the `_PHOENIX_` namespace.